requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
transformers<=4.45.2
accelerate
peft>=0.17.1
opencv-python==4.9.0.80
rospkg==1.5.0
imageio==2.37.0
pillow==11.3.0
diffusers==0.35.1
tensorflow==2.18.0
tf-keras==2.18.0
tensorflow_datasets
tianshou==0.2.0
pytest
dm_control==1.0.14
matplotlib
robosuite==1.4.0
imageio-ffmpeg==0.6.0
mujoco==2.3.7
pynput
numpy==1.26.0
deepspeed==0.14.4
pybullet